fixed generatord loss

+++ src/utils/heatmaps.py
@@ -29,9 +29,9 @@ class HeatMap:
 
         # so we can change type during offset counting
         lat_offset *= 100
-        lat_offset = lat_offset - 1
+        lat_offset = lat_offset.int() - 1
         lon_offset *= 100
-        lon_offset = lon_offset - 1
+        lon_offset = lon_offset.int() - 1
         data[:, 0] = lat_offset
         data[:, 1] = lon_offset
